made the Curl_he2ai() take the port number as an int intead, to avoid lots of typecasts all over

+++ lib/hostip.h
@@ -215,7 +215,7 @@ Curl_addrinfo *Curl_ip2addr(in_addr_t num, char *hostname, int port);
 
 /* [ipv4 only] Curl_he2ai() converts a struct hostent to a Curl_addrinfo chain
    and returns it */
-Curl_addrinfo *Curl_he2ai(struct hostent *, unsigned short port);
+Curl_addrinfo *Curl_he2ai(struct hostent *, int port);
 
 /* relocate a hostent struct */
 void Curl_hostent_relocate(struct hostent *h, long offset);
